Configurator refactoring for scheduler server
