+++ plugin/cmd/kube-scheduler/app/BUILD
@@ -5,11 +5,15 @@ licenses(["notice"])
 load(
     "@io_bazel_rules_go//go:def.bzl",
     "go_library",
+    "go_test",
 )
 
 go_library(
     name = "go_default_library",
-    srcs = ["server.go"],
+    srcs = [
+        "configurator.go",
+        "server.go",
+    ],
     tags = ["automanaged"],
     deps = [
         "//pkg/api:go_default_library",
@@ -51,5 +55,12 @@ filegroup(
         ":package-srcs",
         "//plugin/cmd/kube-scheduler/app/options:all-srcs",
     ],
+    tags = ["automanaged"],
+)
+
+go_test(
+    name = "go_default_test",
+    srcs = ["configurator_test.go"],
+    library = ":go_default_library",
     tags = ["automanaged"],
 )
