+++ plugin/cmd/kube-scheduler/app/configurator_test.go
@@ -1,0 +1,31 @@
+/*
+Copyright 2017 The Kubernetes Authors.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+    http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+
+package app
+
+import (
+	"testing"
+)
+
+func TestSchedulerConfiguratorFailure(t *testing.T) {
+	sc := &schedulerConfigurator{
+	// policyfile and algorithm are intentionally undefined.
+	}
+	_, error := sc.Create()
+	if error == nil {
+		t.Fatalf("Expected error message when creating with incomplete configurator.")
+	}
+}
